pyproject.toml
[project]
name = "sturture_from_motion_project"
version = "1.0.0"
description = "Reconstruct 3D object based on the images"
authors = ["Nguyen Quang Chien <2431101@student.hcmute.edu.vn>"]  
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python >=4.5.1",
    "matplotlib >=3.4.3",
    "python-dotenv >=0.19.0",
]


